simple-rag-chatbot/requirement.txt
chainlit
groq
langchain-community
langchain-core
langchain-text-splitters
faiss-cpu
python-dotenv
pydf
pypdf
pinecone